Factor out RunDependencies component.
