+++ vue/src/views/RunView.vue
@@ -67,81 +67,7 @@ div
               td {{ rec.message }}
 
     Frame(title="Dependencies")
-      div.dependencies
-        div
-        div.l Job
-        div.c Run
-        div.c State
-        div.r Schedule
-        div.r Start
-        div.r Elapsed
-        div.underline1(style="grid-column-end: span 7")
-
-        template(v-if="dependencies !== null")
-          div.colhead.v(:style="{ 'grid-row-end': 'span ' + dependencies.reduce((s, d) => s + Math.max(d.runs.length, 1), 1) }") Dependencies
-          template(v-for="dep, d of dependencies")
-            div.l.col-job(:style="{ 'grid-row-end': 'span ' + dep.runs.length }")
-              div.v: JobWithArgs(:job-id="dep.job_id" :args="dep.args")
-            template(v-for="r, i of dep.runs")
-              div.c.col-run: Run(:run-id="r.run_id")
-              div.c.col-state: State(:state="r.state")
-              div.r.col-schedule-time: Timestamp(:time="r.times.schedule")
-              div.r.col-start-time: Timestamp(v-if="r.times.running" :time="r.times.running")
-              div.r.col-elapsed: RunElapsed(:run="r")
-            div(v-if="dep.runs.length == 0")
-              div.c.col-run(style="grid-column-end: span 6") (no runs)
-              div
-              div
-              div
-              div
-            div(v-if="d < dependencies.length - 1" style="grid-column-end: span 6; border-bottom: 2px dotted #ccc;")
-          div(v-if="dependencies.length == 0" style="grid-column-end: span 6") (no dependencies)
-
-        div.underline1(style="grid-column-end: span 7")
-
-        div.colhead.v: span: b This run
-        div.l.col-job: JobWithArgs(:job-id="run.job_id" :args="run.args")
-        div.c.col-run: Run(:run-id="run_id")
-        div.c.col-state: State(:state="run.state")
-        div.r.col-schedule-time: Timestamp(:time="run.times.schedule")
-        div.r.col-start-time: Timestamp(v-if="run.times.running" :time="run.times.running")
-        div.r.col-elapsed: RunElapsed(:run="run")
-
-        div.underline1(style="grid-column-end: span 7")
-
-        template(v-if="dependents !== null")
-          div.colhead.v(:style="{ 'grid-row-end': 'span ' + dependents.reduce((s, d) => s + Math.max(d.runs.length, 1), 1) }") Dependents
-          template(v-for="dep, d of dependents")
-            div.l.col-job(:style="{ 'grid-row-end': 'span ' + dep.runs.length }")
-              div.v: JobWithArgs(:job-id="dep.job_id" :args="dep.args")
-            template(v-for="r, i of dep.runs")
-              div.c.col-run: Run(:run-id="r.run_id")
-              div.c.col-state: State(:state="r.state")
-              div.r.col-schedule-time: Timestamp(:time="r.times.schedule")
-              div.r.col-start-time: Timestamp(v-if="r.times.running" :time="r.times.running")
-              div.r.col-elapsed: RunElapsed(:run="r")
-            div(v-if="dep.runs.length == 0")
-              div.c.col-run(style="grid-column-end: span 6") (no runs)
-              div
-              div
-              div
-              div
-            div(v-if="d < dependents.length - 1" style="grid-column-end: span 6; border-bottom: 2px dotted #ccc;")
-          div(v-if="dependents.length == 0" style="grid-column-end: span 6") (no dependents)
-
-    Frame(title="Metadata" closed)
-      table.fields
-        tbody
-          tr(v-for="(value, key) in meta && meta.program || {}" :key="key")
-            th {{ key }}
-            td
-              tt(v-if="typeof value === 'object'")
-                table.fields.subfields
-                  tbody
-                    tr(v-for="(sv, sk) in value" :key="sk")
-                      th {{ sk }}
-                      td {{ sv }}
-              tt(v-else) {{ value }}
+      RunDependencies(:run="run")
 
     Frame(title="Output" v-if="hasOutput")
       div.output(v-if="outputMetadata")
@@ -177,8 +103,9 @@ import JobWithArgs from '@/components/JobWithArgs'
 import OperationButton from '@/components/OperationButton'
 import Program from '@/components/Program'
 import Run from '@/components/Run'
-import { getDependencies, getDependents, isComplete, OPERATIONS } from '@/runs'
+import { isComplete, OPERATIONS } from '@/runs'
 import RunArgs from '@/components/RunArgs'
+import RunDependencies from '@/components/RunDependencies'
 import RunElapsed from '@/components/RunElapsed'
 import RunsList from '@/components/RunsList'
 import State from '@/components/State'
@@ -197,6 +124,7 @@ export default {
     Program,
     Run,
     RunArgs,
+    RunDependencies,
     RunElapsed,
     RunsList,
     State,
@@ -243,14 +171,6 @@ export default {
 
     metadata() {
       return this.metadata ? this.metadata.program || {} : {}
-    },
-
-    dependencies() {
-      return this.run ? getDependencies(this.run, this.store) : null
-    },
-
-    dependents() {
-      return this.run ? getDependents(this.run, this.store) : null
     },
   },
 
@@ -432,57 +352,6 @@ export default {
 
 .buttons {
   margin-bottom: 1.5rem;
-}
-
-.dependencies {
-  display: grid;
-  grid-template-columns: repeat(7, max-content);
-  column-gap: 8px;
-  row-gap: 2px;
-  padding: 16px 0;
-
-  > div {
-    margin: 4px 4px;
-  }
-
-  > .l { text-align: left; }
-  > .c { text-align: center; }
-  > .r { justify-self: end; }
-
-  .v {
-    height: 100%;
-    display: flex;
-    align-items: center;
-  }
-
-  .colhead {
-    margin-top: 0;
-    margin-right: 0;
-    padding-right: 16px;
-    border-right: 2px dotted #ccc;
-  }
-
-  .underline1 {
-    border-bottom: 2px solid #ccc;
-  }
-
-  .col-job {
-  }
-
-  .col-run {
-  }
-
-  .col-state {
-  }
-
-  .col-schedule-time, .col-start-time {
-    font-size: 90%;
-    color: $global-light-color;
-  }
-
-  .col-elapsed {
-    white-space: nowrap;
-  }
 }
 
 .output {
